fix: check for file existence before unlinking (#570)
check for file existence before unlinking

+++ src/system/sync/fileCache/clearFileCache.ts
@@ -1,6 +1,9 @@
-import { unlink } from "react-native-fs"
+import { exists, unlink } from "react-native-fs"
 import { PATH_CACHE } from "system/sync/fileCache/constants"
 
 export const clearFileCache = async () => {
-  await unlink(PATH_CACHE)
+  const fileExists = await exists(PATH_CACHE)
+  if (fileExists) {
+    await unlink(PATH_CACHE)
+  }
 }
